requirements.txt
aiogram==3.22.0
sqlalchemy==2.0.23
apscheduler==3.10.4
beautifulsoup4==4.12.2
httpx==0.25.2
playwright==1.40.0
python-dotenv==1.0.0
lxml==4.9.3
